src/pmdb/requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
certifi==2018.8.24
chardet==3.0.4
coverage==4.5.1
decorator==4.3.0
Django==1.11
django-extensions==2.1.3
django-filter==2.0.0
djangorestframework==3.8.2
idna==2.7
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.1
Markdown==3.0.1
more-itertools==4.3.0
parameterized==0.6.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.7.1
prompt-toolkit==2.0.5
ptyprocess==0.6.0
py==1.6.0
Pygments==2.2.0
pytest==3.8.2
pytest-cov==2.6.0
pytest-django==3.4.3
pytz==2018.5
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
